some "working" obstacle avoidance, etc
does what its supposed to do, but poorly.

some weird behavior because the robot's world starts at 0, 0 but the start pos of the sim
is 50, 35. attempts to rectify this have resulted in kinda messy behavior.

need to use the z_angle rotation of the cube's pose in order to better reflect the obstacle
locations in the simulation.

pauses between all movements to 'give it some time to think'. Without doing this, things
tend to go awry. (placing all found cubes at [0, 0] for some reason?) The pauses mostly
resolve this

the conditions for the main loops in cozmo_planning are stupid but w/e its a quick fix

"just needs a bit of fine tuning!" 🙃👍